add shadows
To prevent ugly shadow artefacts due to floating-point inaccuracies, I
slightly move the hit point in normal direction before doing the shadow
test. This way, I prevent that hit points are sometimes slightly below
the surface of an object.

For now, I move by 1e-10 times the normal vector, but this number should
actually depend on the transformation of an object as it contributes to
the inaccuracies in the computation of intersections.