refactor groups
Previously, objects in groups were tested twice for hits, once as part
of the group and once, because we were iterating over all objects in the
hit detection code. To fix this, I've now defined a `scene` vector which
contains all objects (shapes or groups) included in the scene. It does
not include objects that are contained in groups.

Group data, like children and bounding boxes, are now stored in a new
Groups class to prevent a borrowing conflict with two borrows of the
objects vector when adding an object to a group. The Group struct is now
essentially only an index into the vectors contained in a Groups object.

To simplify the creation of groups, I've added some convenience methods
to the World class. In particular the methods `add_group`,
`add_subgroup` and `add_object_to_group`.