fix bug in intersection when using groups
I've forgotten to iterate over the new scene vector instead of the
objects vector in intersection. This resulted in objects in groups
existing twice: once inside their group and once without the group
transoformation applied.

I've added a new unit test to check that objects in groups are correctly
transformed with their group's transformation in the intersect method.
That should catch similar bugs in the future.