add new class ObjectStore
This new class serves as a unified container for objects and groups. For
the moment it simplifies some method signatures that previously took a
&[Object] and a &Groups to only take an &ObjectStore at the cost of
slightly more indirections.

This class will however be useful in the future to implement
constructive solid geometry, which would introduce a new group-like
class CSG. ObjectStore can thus help to resolve Group or CSG index
objects to the right objects.