fix computation of bounding boxes
To enforce correctly computed bounding boxes without recomputing
bounding boxes of all parents when adding a shape, I've split the World
class into two classes: WorldBuilder and World.
The new WorldBuilder class has methods to add new shapes, groups and
CSGs, but no methods for ray intersections. To render a scene, the user
has to convert the WorldBuilder into a World by calling its .build()
method. This consumes the WorldBuilder to create a new World with all
the bounding boxes correctly computed by recursing down the group and
CSG hierarchies.

The World object can then be used for rendering but does not contain any
methods to modify it. This way, we make sure that the precomputed
bounding boxes cannot become out-of-date.
It is however possible to convert a World back into a Worldbuilder with
its .modify() method if one needs to add new objects or modify existing
ones.